+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Fix `Container::mount`, `Container::unmount`
 * *BREAKING* `export` option is no longer available on `ContainerCheckpointOptsBuilder`. In place of the option, there is a new `Container::checkpoint_export` method that exports the tarball as a stream while the `Container::checkpoint` method creates a snapshot on the podman side without exporting it. 
 * *BREAKING* Change stream item of `Container::logs` to `TtyChunk`
+* *BREAKING* Fix typo in Container method name `healtcheck` -> `healthcheck` and fix the endpoint itself.
 
 # 0.5.0
 * Fix some generated models